Optimize block encryption in randomizer
This significantly speeds up the computation of indexes (by a factor of
about 6.5 for 10000 indexes, see benchmarking code).

The optimization works the following way: instead of creating a new
encryptor for every iteration, we use the same encryptor every time.
Since we use ECB, each block is still encrypted independently, and so
there is no change in functionality.

I detected the slowness of the initialization of the encryptor with
py-spy.